Set Up Testing Environment and Databases (#13)
* Set Up Testing Environment and Databases

This commit introduces `RSpec` as the testing framework. While
`Minitest` comes with Rails, the additional features of `RSpec`, namely
the `save_and_open_page` option makes `RSpec` more desirable for this
project.

This commit also switches the database from `sqlte3` to `postgres`.
While this project is unlikely to have a complex schema that needs the
power of Postgres, the ability to create arrays on a field led me to
this choice. There may be other niceties that Postgres gives down the
line.

Finally, the following gems were added:
- `standard` for linting
- `factory-bot-rails` to make testing simpler
- `launchy`, to open screenshots of the current state of your system test in the browser

add factorybot, launchy, and standard; remove unneeded gems

* Lint removal

* Configure hound and more lint removal

* add standard linter to hound

* Trying to get Hound to match Standard rules