Add Seed Data
This commit adds in some basic seed data to make development smoother.

While adding posts, it became clear that the associations between
category/post and article/post were not configured properly.  Rails 5
introduced a validation check on all `belongs_to` associations, so that
was overridden with a `required: false` inclusion. A missing join table
between category and post was also included.

Finally the presence validation on `published_at` was removed from Post,
but the default value and presence validation was left in the migration.