Create Blog List
This commit creates the basic scaffolding for the display of blog
posts by category.  It also includes the creation of a home page,
a category index page, and the aforementioned blog lists on a dedicated
category page.

Additionally, localizations were added as well as system tests.

N.B. After switching from feature tests to system tests, I encountered a
Chromedriver error in CI. It apears that the tests need to run headless
in CI, but it isn't yet clear how to configure them to do so.

This is the error that is popping up in CI:

```
Selenium::WebDriver::Error::UnknownError:
            unknown error: Chrome failed to start: exited abnormally.
              (unknown error: DevToolsActivePort file doesn't exist)
              (The process started from chrome location /usr/bin/google-chrome is no longer running, so ChromeDriver is assuming that Chrome has crashed.)
```