Add `post/show` Page
This commit adds the show page for an individual post. It also updates
the featured post by adding a `most_recent_featured_post` method on
Post and checking for the most recent post on the home page instead of
inside the `featured` partial.

Tests and localizations have also been updated.

Some leftover cruft related to running the specs in headless mode was
also removed as part of this work.