Conditionalize Featured Post
Currently there are three templates for posts: regular post, featured
post, and truncated post.  This commit combines the featured and
truncated posts since it is likely the design for those views will be
the same.  To differentiate between them, a conditional on `featured`
was added.  What this change will be visually has not been decided.
Also, it is possible that I would want to only check if a post was
featured on the home page, but not signal that state on a blog list.

The regular post was left as is as tat design will be different from the
other two templates.